# api_getway/api/handlers/models/post.go
package models

type PostRequest struct {
	Title       string `json:"title"`
	Description string `json:"description"`
	ImgUrl      string `json:"img_url"`
}

type Post struct {
	Id          int64  `json:"id"`
	Title       string `json:"title"`
	Description string `json:"description"`
	ImgUrl      string `json:"img_url"`
	CreatedAt   string `json:"created_at"`
	UpdatedAt   string `json:"updated_at"`
}

type IdPostRequest struct {
	Id int64 `json:"id"`
}

type GetAllPostsRequest struct {
	Limit int64 `json:"limit"`
	Page  int64 `json:"page"`
}

type UpdatePostRequest struct {
	Id          int64  `json:"id"`
	Title       string `json:"title"`
	Description string `json:"description"`
	ImgUrl      string `json:"img_url"`
}

type Posts struct {
	Posts []Post `json:"users"`
}